@@ -23,10 +23,10 @@
     "@angular/platform-browser-dynamic": "^5.2.11",
     "@angular/router": "^5.2.11",
     "bootstrap": "^3.3.7",
-    "core-js": "^2.5.4",
+    "core-js": "^2.5.7",
     "jquery": "^3.3.1",
     "moment": "^2.21.0",
-    "rxjs": "^5.5.8",
+    "rxjs": "^5.5.11",
     "sockjs-client": "^1.1.4",
     "stompjs": "^2.3.3",
     "zone.js": "^0.8.20"
@@ -36,20 +36,20 @@
     "@angular/compiler-cli": "^5.2.11",
     "@angular/language-service": "^5.2.11",
     "@types/jasmine": "2.8.7",
-    "@types/node": "^10.1.1",
+    "@types/node": "^10.1.3",
     "codelyzer": "^4.2.1",
     "jasmine-core": "~3.1.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^2.0.2",
     "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^2.0.0",
+    "karma-coverage-istanbul-reporter": "^2.0.1",
     "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^1.1.0",
     "protractor": "~5.3.2",
     "ts-node": "~3.3.0",
     "tslint": "~5.10.0",
     "typescript": "^2.6.2",
-    "webpack": "^4.8.3"
+    "webpack": "^4.9.1"
   }
 }
